#1 Moved util function to dedicated file
