+++ src/tasks/tasks.go
@@ -4,10 +4,6 @@ import (
 	"encoding/json"
 	"errors"
 	"io/ioutil"
-	"os"
-	"os/user"
-	"strconv"
-	"strings"
 	"time"
 )
 
@@ -127,43 +123,4 @@ func (task Task) getLoggedTime() int {
 	}
 
 	return loggedTime
-}
-
-func check(err error) {
-	if err != nil {
-		panic(err)
-	}
-}
-
-func getCurrentDateTime() time.Time {
-	return time.Now()
-}
-
-func fileExists(path string) bool {
-	_, err := os.Stat(path)
-
-	if err != nil {
-		return false
-	}
-
-	return true
-}
-
-func getTasksDir() string {
-	usr, _ := user.Current()
-
-	return usr.HomeDir + "/.taskctl/tasks"
-}
-
-func getTaskPath(name string) string {
-	name = strings.ToUpper(name)
-
-	return getTasksDir() + "/" + name + ".json"
-}
-
-func formatDuration(seconds int) string {
-	hours := seconds / 3600
-	minutes := seconds % 3600 / 60
-
-	return strconv.Itoa(hours) + " h " + strconv.Itoa(minutes) + " m"
 }
